requirements.txt
django==2.1.15
djangorestframework==3.9.0